+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 * fix: Fix Collectionx's chunked method bug
 * new: Sequencex added chunked, windowed method
 * change: Arrayx, Collectionx, Mapx now need to return empty Lists are Collections.EmptyList
-* fix: Fix the bug that Datex.formatTimeLength() method should return '1590ms' instead of '1ms' when inputting'(1590L, "%ms?ms")' 
+* fix: Fix the bug that Datex.formatTimeLength(1590L, "%ms?ms") should return '1590ms' instead of '1ms'
+* fix: Fix the bug that Datex.formatTimeLength(90L, "%MS?ms") should return '090ms' instead of '90ms'
 
 
 ## v1.0.0-rc01